Bug 1212939 - Disable update logging if given a log file path longer than MAX_PATH; r=rstrong
UltraBlame original commit: 81e422d038d3752790fb8b073faed8351eaf3e16

+++ toolkit/mozapps/update/common/updatelogging.cpp
@@ -28,29 +28,43 @@ void UpdateLog::Init(NS_tchar* sourcePath,
 
 #ifdef XP_WIN
   GetTempFileNameW(sourcePath, L"log", 0, mTmpFilePath);
-  if (append) {
-    NS_tsnprintf(mDstFilePath, sizeof(mDstFilePath)/sizeof(mDstFilePath[0]),
-      NS_T("%s/%s"), sourcePath, alternateFileName);
-    MoveFileExW(mDstFilePath, mTmpFilePath, MOVEFILE_REPLACE_EXISTING);
-  } else {
-    NS_tsnprintf(mDstFilePath, sizeof(mDstFilePath)/sizeof(mDstFilePath[0]),
-                 NS_T("%s/%s"), sourcePath, fileName);
-  }
+  int dstFilePathLen = NS_tsnprintf(mDstFilePath,
+    sizeof(mDstFilePath)/sizeof(mDstFilePath[0]),
+    NS_T("%s/%s"), sourcePath, append ? alternateFileName : fileName);
 
-  logFP = NS_tfopen(mTmpFilePath, append ? NS_T("a") : NS_T("w"));
   
   
-  DeleteFileW(mDstFilePath);
+  if ((dstFilePathLen > 0) &&
+      (dstFilePathLen <
+         static_cast<int>(sizeof(mDstFilePath)/sizeof(mDstFilePath[0])))) {
+    if (append) {
+      MoveFileExW(mDstFilePath, mTmpFilePath, MOVEFILE_REPLACE_EXISTING);
+    }
+    logFP = NS_tfopen(mTmpFilePath, append ? NS_T("a") : NS_T("w"));
+
+    
+    
+    DeleteFileW(mDstFilePath);
+  }
 #else
-  NS_tsnprintf(mDstFilePath, sizeof(mDstFilePath)/sizeof(mDstFilePath[0]),
-               NS_T("%s/%s"), sourcePath, fileName);
-
+  int dstFilePathLen = 0;
   if (alternateFileName && NS_taccess(mDstFilePath, F_OK)) {
-    NS_tsnprintf(mDstFilePath, sizeof(mDstFilePath)/sizeof(mDstFilePath[0]),
+    dstFilePathLen = NS_tsnprintf(mDstFilePath,
+      sizeof(mDstFilePath)/sizeof(mDstFilePath[0]),
       NS_T("%s/%s"), sourcePath, alternateFileName);
+  } else {
+    dstFilePathLen = NS_tsnprintf(mDstFilePath,
+      sizeof(mDstFilePath)/sizeof(mDstFilePath[0]),
+      NS_T("%s/%s"), sourcePath, fileName);
   }
 
-  logFP = NS_tfopen(mDstFilePath, append ? NS_T("a") : NS_T("w"));
+  
+  
+  if ((dstFilePathLen > 0) &&
+      (dstFilePathLen <
+         static_cast<int>(sizeof(mDstFilePath)/sizeof(mDstFilePath[0])))) {
+    logFP = NS_tfopen(mDstFilePath, append ? NS_T("a") : NS_T("w"));
+  }
 #endif
 }
